Move most testing fixtures to re-usable module
